doggy_dates/requirements.txt
alembic==1.0.11
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.1.0
Flask-Bcrypt==0.7.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.17.0
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.5
Werkzeug==0.15.4
